+++ client/src/main/scala/org/http4s/client/middleware/RequestLogger.scala
@@ -23,6 +23,7 @@ import cats.effect.concurrent.Ref
 import cats.syntax.all._
 import fs2._
 import org.log4s.getLogger
+import org.http4s.internal.{Logger => InternalLogger}
 import org.typelevel.ci.CIString
 
 /** Simple Middleware for Logging Requests As They Are Processed
@@ -30,13 +31,21 @@ import org.typelevel.ci.CIString
 object RequestLogger {
   private[this] val logger = getLogger
 
+  private def defaultLogAction[F[_]: Sync](s: String): F[Unit] = Sync[F].delay(logger.info(s))
+
   def apply[F[_]: Concurrent](
       logHeaders: Boolean,
       logBody: Boolean,
       redactHeadersWhen: CIString => Boolean = Headers.SensitiveHeaders.contains,
       logAction: Option[String => F[Unit]] = None
   )(client: Client[F]): Client[F] =
-    impl[F](logHeaders, Left(logBody), redactHeadersWhen, logAction)(client)
+    impl(client, logBody) { request =>
+      Logger.logMessage[F, Request[F]](request)(
+        logHeaders,
+        logBody,
+        redactHeadersWhen
+      )(logAction.getOrElse(defaultLogAction[F]))
+    }
 
   def logBodyText[F[_]: Concurrent](
       logHeaders: Boolean,
@@ -44,36 +53,29 @@ object RequestLogger {
       redactHeadersWhen: CIString => Boolean = Headers.SensitiveHeaders.contains,
       logAction: Option[String => F[Unit]] = None
   )(client: Client[F]): Client[F] =
-    impl[F](logHeaders, Right(logBody), redactHeadersWhen, logAction)(client)
-
-  private def impl[F[_]: Concurrent](
-      logHeaders: Boolean,
-      logBodyText: Either[Boolean, Stream[F, Byte] => Option[F[String]]],
-      redactHeadersWhen: CIString => Boolean,
-      logAction: Option[String => F[Unit]]
-  )(client: Client[F]): Client[F] = {
-    val log = logAction.getOrElse { (s: String) =>
-      Sync[F].delay(logger.info(s))
+    impl(client, logBody = true) { request =>
+      InternalLogger.logMessageWithBodyText[F, Request[F]](request)(
+        logHeaders,
+        logBody,
+        redactHeadersWhen
+      )(logAction.getOrElse(defaultLogAction[F]))
     }
 
-    def logMessage(r: Request[F]): F[Unit] =
-      logBodyText match {
-        case Left(bool) =>
-          Logger.logMessage[F, Request[F]](r)(logHeaders, bool, redactHeadersWhen)(log(_))
-        case Right(f) =>
-          org.http4s.internal.Logger
-            .logMessageWithBodyText[F, Request[F]](r)(logHeaders, f, redactHeadersWhen)(log(_))
-      }
-
-    val logBody: Boolean = logBodyText match {
-      case Left(bool) => bool
-      case Right(_) => true
+  def customized[F[_]: Concurrent](
+      client: Client[F],
+      logBody: Boolean = true,
+      logAction: Option[String => F[Unit]] = None
+  )(requestToText: Request[F] => F[String]): Client[F] =
+    impl(client, logBody) { request =>
+      val log = logAction.getOrElse(defaultLogAction[F] _)
+      requestToText(request).flatMap(log)
     }
 
+  private def impl[F[_]](client: Client[F], logBody: Boolean)(logMessage: Request[F] => F[Unit])(
+      implicit F: Concurrent[F]): Client[F] =
     Client { req =>
       if (!logBody)
-        Resource.liftF(logMessage(req)) *> client
-          .run(req)
+        Resource.liftF(logMessage(req)) *> client.run(req)
       else
         Resource.suspend {
           Ref[F].of(Vector.empty[Chunk[Byte]]).map { vec =>
@@ -87,7 +89,11 @@ object RequestLogger {
                 // Cannot Be Done Asynchronously - Otherwise All Chunks May Not Be Appended Previous to Finalization
                 .observe(_.chunks.flatMap(s => Stream.eval_(vec.update(_ :+ s))))
                 .onFinalizeWeak(
-                  logMessage(req.withBodyStream(newBody))
+                  logMessage(req.withBodyStream(newBody)).attempt
+                    .flatMap {
+                      case Left(t) => F.delay(logger.error(t)("Error logging request body"))
+                      case Right(()) => F.unit
+                    }
                 )
             )
 
@@ -95,6 +101,38 @@ object RequestLogger {
           }
         }
     }
-  }
+
+  val defaultRequestColor: String = Console.BLUE
+
+  def colored[F[_]: Concurrent](
+      logHeaders: Boolean,
+      logBody: Boolean,
+      redactHeadersWhen: CIString => Boolean = Headers.SensitiveHeaders.contains,
+      color: String = defaultRequestColor,
+      logAction: Option[String => F[Unit]] = None
+  )(client: Client[F]): Client[F] =
+    customized(client, logBody, logAction) { request =>
+      import Console._
+      val methodColor =
+        if (request.method.isSafe) color
+        else YELLOW
+
+      val prelude =
+        s"${request.httpVersion} $methodColor${request.method}$RESET$color $BOLD${request.uri}$RESET$color"
+
+      val headers: String =
+        InternalLogger.defaultLogHeaders[F, Request[F]](request)(logHeaders, redactHeadersWhen)
+
+      val bodyText: F[String] =
+        InternalLogger.defaultLogBody[F, Request[F]](request)(logBody) match {
+          case Some(textF) => textF.map(text => s"""body="$text"""")
+          case None => Sync[F].pure("")
+        }
+
+      def spaced(x: String): String = if (x.isEmpty) x else s" $x"
+
+      bodyText
+        .map(body => s"$color$prelude${spaced(headers)}${spaced(body)}$RESET")
+    }
 
 }